Worked on writing domain id in page table. There was some mistake because of differences between linux and hardware logics. Correcting mistakes didn't solve the problem. Even switching all page tables to a different domain and closing kernel domain triggers a crash. I don't think it's an emulation problem. I already try to flush the page tables with no success. Could be a bad understanding of page tables but checked twice already and compare it with /sys/kernel/debug/kernel_page_table.

+++ include/linux/dik/thread.h
@@ -1,5 +1,5 @@
 // This is for debug. Likely to be removed later
-extern void change_stack_back(size_t domain_id, unsigned long stack);
+extern void change_stack_back(size_t domain_id, unsigned int stack);
 
 extern unsigned long set_task_stack_domain_id(size_t domain_id, struct task_struct *task);
 extern void set_task_DACR(size_t new_dacr, struct task_struct *task);
